Move DECam into its own file (out of models/instrument.py)
